refactor: merge AEpiDataEndpoints into EpiDataContext
Collapse the abstract endpoint base and its single subclass into one
concrete EpiDataContext, defined in _endpoints.py.
