+++ epidatpy/_endpoints.py
@@ -4,12 +4,16 @@ import warnings
 from collections.abc import Mapping, Sequence
 from datetime import date
 from typing import (
-    TYPE_CHECKING,
+    Any,
+    Final,
     Literal,
 )
 
 from epiweeks import Week
+from requests import Session
 
+from ._call import EpiDataCall, _request_with_retry
+from ._constants import BASE_URL, CAST_BASE_URL
 from ._covidcast import GeoType, TimeType, define_covidcast_fields
 from ._model import (
     ApiVersion,
@@ -20,14 +24,11 @@ from ._model import (
     EpiRangeParam,
     IntParam,
     InvalidArgumentException,
-    ParamType,
     StringParam,
+    add_endpoint_to_url,
     format_list,
 )
 from ._parse import parse_api_date, parse_user_date_or_week, validate_report_time_query
-
-if TYPE_CHECKING:
-    from .request import EpiDataCall
 
 
 def get_wildcard_equivalent_dates(time_value: EpiRangeParam, time_type: Literal["day", "week"]) -> EpiRangeParam:
@@ -39,19 +40,83 @@ def get_wildcard_equivalent_dates(time_value: EpiRangeParam, time_type: Literal[
     return time_value
 
 
-class AEpiDataEndpoints:
-    """epidata endpoint list and fetcher"""
+class EpiDataContext:
+    """Endpoint catalog and synchronous fetcher for Delphi's Epidata API."""
+
+    _base_url: Final[str]
+    _cast_base_url: Final[str]
+    _session: Final[Session | None]
+
+    def __init__(
+        self,
+        base_url: str = BASE_URL,
+        session: Session | None = None,
+        use_cache: bool | None = None,
+        cache_max_age_days: int | None = None,
+        cast_base_url: str = CAST_BASE_URL,
+    ) -> None:
+        self._base_url = base_url
+        self._cast_base_url = cast_base_url
+        self._session = session
+        self.use_cache = use_cache
+        self.cache_max_age_days = cache_max_age_days
+
+    def with_base_url(self, base_url: str) -> EpiDataContext:
+        return EpiDataContext(
+            base_url,
+            self._session,
+            self.use_cache,
+            self.cache_max_age_days,
+            cast_base_url=self._cast_base_url,
+        )
+
+    def with_session(self, session: Session) -> EpiDataContext:
+        return EpiDataContext(
+            self._base_url,
+            session,
+            self.use_cache,
+            self.cache_max_age_days,
+            cast_base_url=self._cast_base_url,
+        )
 
     def _create_call(
         self,
         endpoint: str,
-        params: Mapping[str, ParamType | None],
+        params: Mapping[str, EpiRangeParam | None],
         meta: Sequence[EpidataFieldInfo] | None = None,
         only_supports_classic: bool = False,
         api_version: ApiVersion = "classic",
         post_filter: CastPostFilter | None = None,
     ) -> EpiDataCall:
-        raise NotImplementedError()
+        base_url = self._cast_base_url if api_version == "cast" else self._base_url
+        return EpiDataCall(
+            base_url,
+            self._session,
+            endpoint,
+            params,
+            meta,
+            only_supports_classic,
+            self.use_cache,
+            self.cache_max_age_days,
+            api_version=api_version,
+            post_filter=post_filter,
+        )
+
+    def epidata_meta(self, source: str) -> Any:
+        """Fetch source-level metadata from the CAST API.
+
+        Returns the parsed JSON (a list of signal/geo descriptors) for `source`.
+        """
+        url = add_endpoint_to_url(self._cast_base_url, "metadata/")
+        response = _request_with_retry(
+            url,
+            {"source": source},
+            self._session,
+            stream=False,
+            api_version="cast",
+        )
+        response.raise_for_status()
+        return response.json()
 
     def pvt_cdc(
         self,
